refactor common HtmlWebpackPlugin implementation
